Move: attraction data
move attraction data from infra.tourapi.vo to attraction.data
