+++ src/main/java/org/routemaster/api/total/domain/attraction/controller/DetailSearchRestController.java
@@ -9,8 +9,8 @@ import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.routemaster.api.total.domain.attraction.service.AttractionDetailSearchService;
-import org.routemaster.api.total.infra.tourapi.vo.AttractionSearchVO;
-import org.routemaster.api.total.infra.tourapi.vo.detail.*;
+import org.routemaster.api.total.domain.attraction.data.search.AttractionSearchVO;
+import org.routemaster.api.total.domain.attraction.data.detail.*;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
