+++ src/main/java/org/routemaster/api/total/domain/attraction/controller/StaySearchRestController.java
@@ -8,8 +8,8 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.routemaster.api.total.domain.attraction.data.search.AttractionSearchVO;
 import org.routemaster.api.total.domain.attraction.service.AttractionSearchService;
-import org.routemaster.api.total.infra.tourapi.vo.AttractionSearchVO;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
