+++ src/main/java/org/routemaster/api/total/domain/attraction/data/detail/TourAttractionDetailVO.java
@@ -1,4 +1,4 @@
-package org.routemaster.api.total.infra.tourapi.vo.detail;
+package org.routemaster.api.total.domain.attraction.data.detail;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import io.swagger.v3.oas.annotations.media.Schema;
