+++ src/main/java/org/routemaster/api/total/domain/attraction/data/search/AreaBasedAttractionVO.java
@@ -1,4 +1,4 @@
-package org.routemaster.api.total.infra.tourapi.vo.attraction;
+package org.routemaster.api.total.domain.attraction.data.search;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import io.swagger.v3.oas.annotations.media.Schema;
