+++ src/main/java/org/routemaster/api/total/domain/attraction/data/search/AttractionSearchVO.java
@@ -1,11 +1,11 @@
-package org.routemaster.api.total.infra.tourapi.vo;
+package org.routemaster.api.total.domain.attraction.data.search;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.*;
 import lombok.extern.slf4j.Slf4j;
-import org.routemaster.api.total.infra.tourapi.vo.attraction.*;
-import org.routemaster.api.total.infra.tourapi.vo.detail.CommonDetailVO;
+import org.routemaster.api.total.domain.attraction.data.detail.CommonDetailVO;
+import org.routemaster.api.total.domain.attraction.data.search.*;
 
 import java.util.ArrayList;
 import java.util.List;
