+++ src/main/java/org/routemaster/api/total/domain/attraction/data/search/StayAttractionVO.java
@@ -1,9 +1,8 @@
-package org.routemaster.api.total.infra.tourapi.vo.attraction;
+package org.routemaster.api.total.domain.attraction.data.search;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
-import org.routemaster.api.total.infra.tourapi.vo.attraction.AttractionVO;
 
 import java.text.ParseException;
