+++ src/main/java/org/routemaster/api/total/domain/attraction/service/AttractionSearchService.java
@@ -1,6 +1,6 @@
 package org.routemaster.api.total.domain.attraction.service;
 
-import org.routemaster.api.total.infra.tourapi.vo.AttractionSearchVO;
+import org.routemaster.api.total.domain.attraction.data.search.AttractionSearchVO;
 import reactor.core.publisher.Mono;
 
 public interface AttractionSearchService {
