+++ src/main/java/org/routemaster/api/total/domain/attraction/service/impl/DefaultAttractionDetailSearchService.java
@@ -5,8 +5,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.extern.slf4j.Slf4j;
 import org.routemaster.api.total.domain.attraction.service.AttractionDetailSearchService;
 import org.routemaster.api.total.infra.tourapi.value.TourAPI;
-import org.routemaster.api.total.infra.tourapi.vo.AttractionSearchVO;
-import org.routemaster.api.total.infra.tourapi.vo.detail.*;
+import org.routemaster.api.total.domain.attraction.data.search.AttractionSearchVO;
+import org.routemaster.api.total.domain.attraction.data.detail.*;
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Service;
 import org.springframework.web.reactive.function.client.WebClient;
